+++ benchmark/graph_color/cs.jl
@@ -1,8 +1,9 @@
-using ConstraintSolver, JuMP
+using ConstraintSolver, JuMP, GLPK
 CS = ConstraintSolver
 
 function main(filename; benchmark = false, time_limit=100)
-    m = Model(optimizer_with_attributes(CS.Optimizer, "time_limit"=>time_limit))
+    lp_optimizer = optimizer_with_attributes(GLPK.Optimizer, "msg_lev" => GLPK.GLP_MSG_OFF)
+    m = Model(optimizer_with_attributes(CS.Optimizer, "time_limit"=>time_limit, "lp_optimizer" => lp_optimizer))
 
     lines = readlines(filename)
     num_colors = 0
@@ -10,7 +11,8 @@ function main(filename; benchmark = false, time_limit=100)
     max_color = nothing
     degrees = nothing
     for line in lines
-        parts = split(line, " ")
+        isempty(line) && continue
+        parts = split(line)
         if parts[1] == "p"
             num_colors = parse(Int, parts[3])
             @variable(m, 1 <= max_color <= num_colors, Int)
@@ -36,6 +38,9 @@ function main(filename; benchmark = false, time_limit=100)
     optimize!(m)
 
     status = JuMP.termination_status(m)
-
-    print("$status, $(JuMP.objective_value(m)), $(JuMP.solve_time(m))")
+    if status == MOI.OPTIMAL
+        print("$status, $(JuMP.objective_value(m)), $(JuMP.solve_time(m))")
+    else
+        print("$status, NaN, $(time_limit)")
+    end
 end