+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # ConstrainSolver.jl - Changelog
 
-## v0.3.1
+## Unreleased
+- Use LP solver also for single variable objectives
+- Combine several `x != y` constraints into all different constraints
+- Combine several `a >= x` constraints with the same `a` into a vector constraints
+  - used for better bounds using all different constraints
+
+## v0.3.1 (16th of November 2020)
 - Added `copy` function for constraint structs for latest JuMP/MOI versions
 
-## v0.3.0
+## v0.3.0 (11th of July 2020)
 - Reified constraint [#171](https://github.com/Wikunia/ConstraintSolver.jl/pull/171)
 
 ## v0.2.2 (26th of June 2020)